Add order endpoints

+++ src/orders/orders.controller.ts
@@ -1,5 +1,16 @@
-import { Controller, Get, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Delete,
+  Post,
+  Param,
+  Body,
+  Put,
+} from '@nestjs/common';
 import { OrdersService } from './orders.service';
+import { ParseUUIDPipe, NotFoundException } from '@nestjs/common';
+import { CreateOrderDTO } from './dtos/create-order.dto';
+import { UpdateOrderDTO } from './dtos/update-order.dto';
 
 @Controller('orders')
 export class OrdersController {
@@ -11,13 +22,31 @@ export class OrdersController {
   }
 
   @Get('/:id')
-  getById(@Param('id') id: string) {
-    return this.ordersService.getById(id);
+  getById(@Param('id', new ParseUUIDPipe()) id: string) {
+    const ord = this.ordersService.getById(id);
+    if (!ord) throw new NotFoundException('Order not found');
+    return ord;
   }
 
   @Delete('/:id')
-  deleteById(@Param('id') id: string) {
+  deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.ordersService.getById(id))
+      throw new NotFoundException('Order not found');
     this.ordersService.deleteById(id);
+    return { success: true };
+  }
+
+  @Post('/')
+  create(@Body() orderData: CreateOrderDTO) {
+    return this.ordersService.create(orderData);
+  }
+
+  @Put('/:id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() orderData: UpdateOrderDTO,
+  ) {
+    this.ordersService.updateById(id, orderData);
     return { success: true };
   }
 }
