+++ src/products/products.controller.ts
@@ -1,7 +1,16 @@
-import { Controller, Get, Delete, Post, Param, Body } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Delete,
+  Post,
+  Param,
+  Body,
+  Put,
+} from '@nestjs/common';
 import { ProductsService } from './products.service';
-
+import { ParseUUIDPipe, NotFoundException } from '@nestjs/common';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -13,12 +22,16 @@ export class ProductsController {
   }
 
   @Get('/:id')
-  getById(@Param('id') id: string) {
-    return this.productsService.getById(id);
+  getById(@Param('id', new ParseUUIDPipe()) id: string) {
+    const prod = this.productsService.getById(id);
+    if (!prod) throw new NotFoundException('Product not found');
+    return prod;
   }
 
   @Delete('/:id')
-  deleteById(@Param('id') id: string) {
+  deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
     this.productsService.deleteById(id);
     return { success: true };
   }
@@ -26,5 +39,14 @@ export class ProductsController {
   @Post('/')
   create(@Body() ProductData: CreateProductDTO) {
     return this.productsService.create(ProductData);
+  }
+
+  @Put('/:id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+    this.productsService.updateById(id, productData);
+    return { success: true };
   }
 }
